+++ packages/web/app/views/administration/users/index.js
@@ -1,2 +1,3 @@
 export { UserAdminView } from './UserAdminView';
 export { UserProfileModal } from './UserProfileModal';
+export { RolesAndDesignationsAdminView } from './RolesAndDesignationsAdminView';
